+++ sample/index.php
@@ -384,6 +384,36 @@ if (PHP_SAPI == 'cli') {
                             </div>
                         </div>
                     </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8"><h5>Get transaction confidence</h5></div>
+                            <div class="col-md-4">
+                                <a href="transaction-confidence/GetTransactionConfidence.php"
+                                   class="btn btn-primary pull-left execute"> Try
+                                    It <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/transaction-confidence/GetTransactionConfidence.html"
+                                   class="btn btn-default pull-right">Source <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8"><h5>Get multiple transactions confidence at once</h5></div>
+                            <div class="col-md-4">
+                                <a href="transaction-confidence/GetMultipleTransactionsConfidence.php"
+                                   class="btn btn-primary pull-left execute">
+                                    Try It
+                                    <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/transaction-confidence/GetMultipleTransactionsConfidence.html"
+                                   class="btn btn-default pull-right">Source
+                                    <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
                 </ul>
             </div>